+++ lib/magicwrite.rb
@@ -17,7 +17,7 @@ module MagicWrite
 
   class Configuration
     attr_writer :access_token
-    attr_accessor :uri_base, :request_timeout
+    attr_accessor :uri_base, :request_timeout, :extra_headers
 
     DEFAULT_URI_BASE = 'https://api.magicwrite.ai/rest/'.freeze
     DEFAULT_REQUEST_TIMEOUT = 120
@@ -26,6 +26,7 @@ module MagicWrite
       @access_token = nil
       @uri_base = DEFAULT_URI_BASE
       @request_timeout = DEFAULT_REQUEST_TIMEOUT
+      @extra_headers = nil
     end
 
     def access_token
@@ -38,13 +39,13 @@ module MagicWrite
 
   class << self
     attr_writer :configuration
-  end
 
-  def self.configuration
-    @configuration ||= MagicWrite::Configuration.new
-  end
+    def configuration
+      @configuration ||= MagicWrite::Configuration.new
+    end
 
-  def self.configure
-    yield(configuration)
+    def configure
+      yield(configuration)
+    end
   end
 end