+++ CHANGELOG.md
@@ -5,6 +5,14 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [0.1.5] - 2023-08-22
+
+### Added
+
+- More specs to cover more implementations
+- Refactor client and methods to use client in initalization
+- include HTTP instead of extend for Client class
+
 ## [0.1.4] - 2023-08-22
 
 ### Added